requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
redis
bcrypt
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
aiofiles
httpx 